# apps/web/apps/web-antd/src/api/generated/infra.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RustSet Infra API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/infra/business-application/page": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "pageNo",
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "default": null
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "default": null
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_for_Page_for_AnyValue"
                }
              }
            }
          }
        }
      }
    },
...
    "/infra/business-application/create": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBusinessApplicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_for_string"
                }
              }
            }
          }
        }
      }
    },
    "/infra/business-application/update": {
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBusinessApplicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_for_null"
                }
              }
            }
          }
        }
      }
    },
...
    "/infra/application-endpoint/page": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "pageNo",
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "default": null
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "default": null
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_for_Page_for_AnyValue"
                }
              }
            }
          }
        }
      }
    },
...
    "/infra/application-endpoint/create": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationEndpointRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_for_string"
                }
              }
            }
          }
        }
      }
    },
    "/infra/application-endpoint/update": {
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApplicationEndpointRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_for_null"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_for_Page_for_AnyValue": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "uint16",
            "maximum": 65535,
            "minimum": 0
          },
          "data": {
            "$ref": "#/components/schemas/Page_for_AnyValue"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "data",
          "message"
        ]
      },
...
      "ApiResponse_for_null": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "uint16",
            "maximum": 65535,
            "minimum": 0
          },
          "data": {
            "type": "null"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "data",
          "message"
        ]
      },
      "ApiResponse_for_string": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "uint16",
            "maximum": 65535,
            "minimum": 0
          },
          "data": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "data",
          "message"
        ]
      },
...
      "CreateApplicationEndpointRequest": {
        "type": "object",
        "properties": {
          "businessApplicationId": {
            "type": "integer",
            "format": "int64"
          },
          "createdBy": {
            "type": [
              "string",
              "null"
            ]
          },
          "destIp": {
            "type": "string"
          },
          "destPort": {
            "type": "string"
          },
          "domain": {
            "type": [
              "string",
              "null"
            ]
          },
          "natIp": {
            "type": [
              "string",
              "null"
            ]
          },
          "protocol": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "required": [
          "businessApplicationId",
          "protocol",
          "destIp",
          "destPort"
        ]
      },
...
      "CreateBusinessApplicationRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdBy": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "required": [
          "name"
        ]
      },
...
      "Page_for_AnyValue": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "items": true
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "list",
          "total"
        ]
      },
...
      "UpdateApplicationEndpointRequest": {
        "type": "object",
        "properties": {
          "businessApplicationId": {
            "type": "integer",
            "format": "int64"
          },
          "createdBy": {
            "type": [
              "string",
              "null"
            ]
          },
          "destIp": {
            "type": "string"
          },
          "destPort": {
            "type": "string"
          },
          "domain": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "natIp": {
            "type": [
              "string",
              "null"
            ]
          },
          "protocol": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "required": [
          "id",
          "businessApplicationId",
          "protocol",
          "destIp",
          "destPort"
        ]
      },
...
      "UpdateBusinessApplicationRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdBy": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "required": [
          "id",
          "name"
        ]
      },
...
    }
  }
}